重构

+++ ClusteringAlgorithm/ClusteringAlgorithm/Containers/IEnumerableExtension.cs
@@ -1,12 +1,13 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 
 namespace ClusteringAlgorithm.Containers {
     public static class EnumerableExtension {
         /// <summary>
-        ///     计算序列的均值
+        ///     根据加法和除法的定义计算序列的均值
         /// </summary>
         /// <typeparam name="T"></typeparam>
         /// <param name="source"></param>
@@ -15,6 +16,6 @@ namespace ClusteringAlgorithm.Containers {
         /// <returns></returns>
         [SuppressMessage("ReSharper", "PossibleMultipleEnumeration")]
         public static T Average<T>(this IEnumerable<T> source, Func<T, T, T> sumFunc,
-            Func<T, double, T> divFunc) => divFunc(source.Aggregate(sumFunc), source.Count());
+            Func<T, int, T> divFunc) => divFunc(source.Aggregate(sumFunc), source.Count());
     }
 }
